Rename server/zod to server/validators; Convert absolute paths to relative on server build
