+++ server/package.json
@@ -53,6 +53,7 @@
     "node": "^21.7.1",
     "reactflow": "^11.11.3",
     "ts-node": "^10.9.2",
+    "tsc-alias": "^1.8.10",
     "zod": "^3.23.8"
   }
 }
